Allow env variable name to be different from account name
This commit adds an "account" field to the people object in the +page.svelte file. This field is extracted from the split string and assigned as a property of each person object. The purpose of this change is to use the "account" property instead of the "name" property when searching for transactions related to each person.